+++ NEWS.md
@@ -1,5 +1,9 @@
 # tidyr 0.7.2.9000
 
+* `complete()` with zero-length completions returns original input (#331).
+
+* `expand()` with empty input gives empty data frame instead of `NULL` (#331).
+
 * `uncount()` performs the inverse operation of `dplyr::count()` (#279)
 
 * `spread()` now consistently returns 0 row outputs for 0 row inputs (#269).